+++ packages/charts/src/chart_types/heatmap/state/selectors/get_debug_state.ts
@@ -14,15 +14,14 @@ import { Position } from '../../../../utils/common';
 import { computeLegendSelector } from './compute_legend';
 import { geometries } from './geometries';
 import { getHighlightedAreaSelector, getHighlightedDataSelector } from './get_highlighted_area';
-import { getPickedCells } from './get_picked_cells';
 
 /**
  * Returns a stringified version of the `debugState`
  * @internal
  */
 export const getDebugStateSelector = createCustomCachedSelector(
-  [geometries, computeLegendSelector, getHighlightedAreaSelector, getPickedCells, getHighlightedDataSelector],
-  (geoms, legend, pickedArea, pickedCells, highlightedData): DebugState => {
+  [geometries, computeLegendSelector, getHighlightedAreaSelector, getHighlightedDataSelector],
+  (geoms, legend, pickedArea, highlightedData): DebugState => {
     return {
       // Common debug state
       legend: getLegendState(legend),
@@ -50,12 +49,13 @@ export const getDebugStateSelector = createCustomCachedSelector(
       },
       // Heatmap debug state
       heatmap: {
-        cells: geoms.heatmapViewModel.cells.map(({ x, y, fill, formatted, value }) => ({
-          x,
-          y,
-          fill: RGBATupleToString(fill.color),
-          formatted,
-          value,
+        cells: geoms.heatmapViewModel.cells.map((cell) => ({
+          x: cell.x,
+          y: cell.y,
+          fill: RGBATupleToString(cell.fill.color),
+          formatted: cell.formatted,
+          value: cell.value,
+          valueShown: geoms.config.cell.label.visible && Number.isFinite(geoms.heatmapViewModel.cellFontSize(cell)),
         })),
         selection: {
           area: pickedArea,