+++ packages/charts/src/index.ts
@@ -14,7 +14,20 @@ export { SpecId, GroupId, AxisId, AnnotationId } from './utils/ids';
 
 // Everything related to the specs types and react-components
 export * from './specs';
-export { DebugState } from './state/types';
+export {
+  DebugState,
+  DebugStateLine,
+  DebugStateValue,
+  DebugStateAnnotations,
+  DebugStateArea,
+  DebugStateAxes,
+  DebugStateBar,
+  DebugStateLegend,
+  DebugStateLineConfig,
+  DebugStateAxis,
+  DebugStateBase,
+  DebugStateLegendItem,
+} from './state/types';
 export { toEntries } from './utils/common';
 export { CurveType } from './utils/curves';
 export { ContinuousDomain, OrdinalDomain } from './utils/domain';
@@ -62,9 +75,12 @@ export { DARK_THEME } from './utils/themes/dark_theme';
 // partition
 export * from './chart_types/partition_chart/layout/types/viewmodel_types';
 export * from './chart_types/partition_chart/layout/utils/group_by_rollup';
+export { AnimKeyframe } from './chart_types/partition_chart/layout/types/config_types';
 
 // heatmap
 export { Cell } from './chart_types/heatmap/layout/types/viewmodel_types';
+export { SizeRatio, TimeMs } from './common/geometry';
+export { TextAlign, TextBaseline } from './common/text_utils';
 export { Config as HeatmapConfig } from './chart_types/heatmap/layout/types/config_types';
 export { ColorBand, HeatmapBandsColorScale } from './chart_types/heatmap/specs/heatmap';
 
@@ -96,3 +112,14 @@ export { Pixels, Ratio } from './common/geometry';
 export { AdditiveNumber } from './utils/accessor';
 export { FontStyle, FONT_STYLES } from './common/text_utils';
 export { Color } from './common/colors';
+
+export {
+  ESCalendarInterval,
+  ESCalendarIntervalUnit,
+  ESFixedInterval,
+  ESFixedIntervalUnit,
+} from './utils/chrono/elasticsearch';
+
+// data utils
+export { GroupKeysOrKeyFn, GroupByKeyFn } from './chart_types/xy_chart/utils/group_data_series';
+export { computeRatioByGroups } from './utils/data/data_processing';